requirements.txt
Flask == 2.1.1
uuid == 1.30
Flask-SQLAlchemy == 2.5.1
Flask-Migrate == 3.1.0
Flask-RESTful == 0.3.9
PyJWT == 2.3.0
